refactor: 스키마 lib폴더에 정의
