+++ src/pages/_app.tsx
@@ -9,7 +9,7 @@ import ModalProvider from '@/providers/modal-provider';
 import { Toaster } from '@/components/shared/ui/toaster';
 import { GlobalBoundary, Portal, Floating } from '@/components/shared';
 
-import 'public/styles/globals.css';
+import '@/styles/globals.css';
 
 declare global {
   interface Window {
