+++ src/pages/signin.tsx
@@ -3,8 +3,8 @@ import Link from 'next/link';
 import axios from 'axios';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { z } from 'zod';
 
+import { signInSchema } from '@/lib/validators/authValidator';
 import useSignIn from '@/models/auth/useSignIn';
 
 import MogazoaLayout from '@/components/layout/App/MogazoaLayout';
@@ -16,14 +16,6 @@ export interface ILoginForm {
   email: string;
   password: string;
 }
-
-const signInSchema = z.object({
-  email: z
-    .string()
-    .min(1, '이메일을 입력해 주세요.')
-    .email('올바른 이메일 주소가 아닙니다.'),
-  password: z.string().min(8, '비밀번호는 최소 8자 이상이어야 합니다.'),
-});
 
 const SignIn = () => {
   const { mutate } = useSignIn();
